% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westfall_young_stepdown.R
\name{boot_stepdown}
\alias{boot_stepdown}
\title{Free step-down adjusted p-values}
\usage{
boot_stepdown(full_formulas, null_formulas, data, coef_list,
  weights = NULL, cluster = NULL, nboots = 10000,
  boot_type = c("wild", "pairs"), parallel = TRUE, pb = TRUE)
}
\arguments{
\item{full_formulas}{A list of objects of class \code{formula}, specifying
the full (unrestricted) models.}

\item{null_formulas}{A list of objects of class \code{formula}, specifying
the null (restricted) models.}

\item{data}{A data frame containing the variables in the models.}

\item{coef_list}{A character vector specifying the variable for each model
for which the p-value is to be adjusted.}

\item{weights}{An optional list of weights to be used in the fitting process.
Each element of the list should be a numeric vector with length equals to
the number of rows in \code{data}. If \code{NULL}, each observation is
given a weight of 1. Weights can be used to specify subgroups, by giving units outside the subgroup of interest a weight of 0.}

\item{cluster}{An optional character string indicating the column in the
data frame that records the cluster to which an observation belongs (for
computing standard errors that account for clustering).}

\item{nboots}{The number of bootstrap replicates, a single positive integer.}

\item{boot_type}{A character string indicating the type of resampling
required. Possible values are \code{wild} or \code{pairs}.}

\item{parallel}{logical, indicating if parallel operation is to be used. If
\code{TRUE}, a parallel backend should be registered prior to running
\code{boot_stepdown}. See the documentation for \code{foreach} in the
\pkg{foreach} package for more details.}

\item{pb}{logical. Should a progress bar be displayed if \code{parallel} is
\code{FALSE}? Ignored if \code{parallel} is \code{TRUE}.}
}
\value{
A data frame reporting unadjusted and adjusted p-values for each
  hypothesis provided in \code{full_formulas} and \code{null_formulas}.
}
\description{
Generates adjusted p-values using the algorithm described in Westfall and
  Young (1993).
}
\examples{
# Replicate Casey et al 2012, Table II Column 3
F <- lapply(sprintf("zscore_\%d ~ 0 + t + tothhs + road + ward", 1:12),
            as.formula)
pvals <- boot_stepdown(full_formulas = F,
                       null_formulas = lapply(F, update, . ~ . - t),
                       data = gobifo,
                       coef_list = "t",
                       nboots = 100, # small nboots for demonstration only
                       parallel = FALSE,
                       boot_type = "pairs")
dplyr::mutate_if(pvals, is.numeric, round, 3)
}
\references{
Westfall, Peter H., and S. Stanley Young. 1993. \emph{Resampling-
Based Multiple Testing: Examples and Methods for P-Value Adjustment.}
John Wiley & Sons, New York.
}
